Bug fixes
Fixed an issue where the check for time between led value incriments didn't work correctly. Previously the value would overflow since the const incr_time was declared as a int and the mult would then be done with an int as an output. Fixed by changing the constant incr_time to an unsigned long